todo-api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mariadb
python-dotenv